Add RenameField operator
Stil has some kinks to be worked out, as does, I now realize,
RemoveField. The problem is that you need to look up names during
evaluation, but if those names change during lowering, execution can't
look up the old values when executing the expressions from before the
remove.
Not sure what to do about this; we'll have to think about it. For now,
basic use cases work (as demonstrated in tests).